Move remaining tests into src folder and add ewatercycle.testing module
This triggers circular import issues
