+++ src/ewatercycle/testing/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for testing eWaterCycle functionality."""
