+++ tests/plugins/lisflood/test_model.py
@@ -14,7 +14,7 @@ from ewatercycle.forcing import load_foreign
 from ewatercycle.parameter_sets import ParameterSet, example_parameter_sets
 from ewatercycle.parametersetdb.config import XmlConfig
 from ewatercycle.plugins.lisflood.model import Lisflood
-from tests.models.fake_models import FailingModel
+from ewatercycle.testing.fake_models import FailingModel
 
 
 @pytest.fixture
